refactor: reorganize compile process and file structure
