+++ src/core/compile.ts
@@ -1,10 +1,10 @@
 import { GoslingSpec, TemplateTrackDef } from './gosling.schema';
-import { compileLayout } from './layout/layout';
 import { HiGlassSpec } from './higlass.schema';
 import { traverseToFixSpecDownstream, overrideDataTemplates } from './utils/spec-preprocess';
 import { replaceTrackTemplates } from './utils/template';
-import { Size } from './utils/bounding-box';
+import { getRelativeTrackInfo, Size } from './utils/bounding-box';
 import { CompleteThemeDeep } from './utils/theme';
+import { renderHiGlass } from './create-higlass-models';
 
 export function compile(
     spec: GoslingSpec,
@@ -13,17 +13,20 @@ export function compile(
     theme: Required<CompleteThemeDeep>
 ) {
     // Make sure to keep the original spec as is
-    const _spec = JSON.parse(JSON.stringify(spec));
+    const specCopy = JSON.parse(JSON.stringify(spec));
 
     // Override default visual encoding (i.e., `DataTrack` => `BasicSingleTrack`)
-    overrideDataTemplates(_spec);
+    overrideDataTemplates(specCopy);
 
     // Replace track templates with raw gosling specs (i.e., `TemplateTrack` => `SingleTrack | OverlaidTrack`)
-    replaceTrackTemplates(_spec, templates);
+    replaceTrackTemplates(specCopy, templates);
 
     // Fix track specs by looking into the root-level spec
-    traverseToFixSpecDownstream(_spec);
+    traverseToFixSpecDownstream(specCopy);
+
+    // Generate arrangement data
+    const trackInfo = getRelativeTrackInfo(specCopy, theme);
 
     // Make HiGlass models for individual tracks
-    compileLayout(_spec, setHg, theme);
+    renderHiGlass(specCopy, trackInfo, setHg, theme);
 }
