+++ src/core/create-higlass-models.ts
@@ -1,10 +1,10 @@
-import { getBoundingBox, Size, TrackInfo } from '../utils/bounding-box';
-import { goslingToHiGlass } from '../gosling-to-higlass';
-import { HiGlassModel } from '../higlass-model';
-import { HiGlassSpec } from '../higlass.schema';
-import { getLinkingInfo } from '../utils/linking';
-import { GoslingSpec } from '../gosling.schema';
-import { CompleteThemeDeep } from '../utils/theme';
+import { getBoundingBox, Size, TrackInfo } from './utils/bounding-box';
+import { goslingToHiGlass } from './gosling-to-higlass';
+import { HiGlassModel } from './higlass-model';
+import { HiGlassSpec } from './higlass.schema';
+import { getLinkingInfo } from './utils/linking';
+import { GoslingSpec } from './gosling.schema';
+import { CompleteThemeDeep } from './utils/theme';
 
 export function renderHiGlass(
     spec: GoslingSpec,
