+++ src/core/utils/bounding-box.test.ts
@@ -1,5 +1,5 @@
 import { GoslingSpec, Track } from '../gosling.schema';
-import { DEFAULT_CIRCULAR_VIEW_PADDING, DEFAULT_VIEW_SPACING } from '../layout/defaults';
+import { DEFAULT_CIRCULAR_VIEW_PADDING, DEFAULT_VIEW_SPACING } from '../defaults';
 import { getBoundingBox, getRelativeTrackInfo } from './bounding-box';
 import { traverseToFixSpecDownstream } from './spec-preprocess';
 import { getTheme } from './theme';
